split GaloisGrp.jl and breakout Groups, POSet
there is more potential, but this is a start
