+++ experimental/Experimental.jl
@@ -1,0 +1,3 @@
+include("GaloisGrp.jl")
+include("Rings.jl")
+include("ModStd.jl")
