+++ experimental/GaloisGrp/GaloisGrp.jl
@@ -1,4 +1,3 @@
-
 module GaloisGrp
 
 using Oscar, Markdown
@@ -203,24 +202,6 @@ function Nemo.roots_upper_bound(f::fmpq_poly)
   return max(fmpz(1), maximum([ceil(fmpz, abs(coeff(f, i)//a)) for i=0:degree(f)]))
 end
 
-#TODO: check where this should be done properly
-# not needed for Galois stuff.
-function (f::RelSeriesElem)(a::RingElem)
-  y = a
-  v = valuation(f)
-  p = precision(f)
-  z = zero(y)
-  for i = Int(p):-1:Int(v)
-      z *= y
-      c = coeff(f, i)
-      if !iszero(c)
-          z += c
-      end
-  end
-  z *= y^Int(v)
-  return z
-end
-
 #roots are sums of m distinct roots of f
 #from https://doi.org/10.2307/2153295
 #Symmetric Functions, m-Sets, and Galois Groups
@@ -350,52 +331,6 @@ function Hecke.orbit(G::Oscar.PermGroup, f::MPolyElem)
   return s
 end
 
-"""
-Tests if `U` is a maximal subgroup of `G`. Suboptimal algorithm...
-  Missing in GAP
-"""
-function ismaximal(G::Oscar.PermGroup, U::Oscar.PermGroup)
-  if !issubgroup(G, U)[1]
-    return false
-  end
-  if order(G)//order(U) < 100
-    t = right_transversal(G, U)[2:end] #drop the identity
-    if any(x->order(sub(G, vcat(gens(U), [x]))[1]) < order(G), t)
-      return false
-    end
-    return true
-  end
-  S = maximal_subgroups(G)
-  if any(x->x == U, S)
-    return true
-  end
-  return false
-end
-
-"""
-Tests if a conjugate of `V` by some element in `G` is a subgroup of `U`.
-  Missing from GAP
-"""
-function isconjugate_subgroup(G::Oscar.PermGroup, U::Oscar.PermGroup, V::Oscar.PermGroup)
-  if order(V) == 1
-    return true, one(U)
-  end
-  local sigma
-  while true
-    sigma = rand(V)
-    if order(sigma) > 1
-      break
-    end
-  end
-  s = short_right_transversal(G, U, sigma)
-  for t = s
-    if issubgroup(U, V^inv(t))[1]
-      return true, inv(t)
-    end
-  end
-  return false, one(U)
-end
-
 probable_orbit(G::Oscar.PermGroup, f::MPolyElem) = orbit(G, f)
 
 #one cannot compare (==) slpoly, no hash either..
@@ -425,99 +360,12 @@ function probable_orbit(G::Oscar.PermGroup, f::SLPoly)
   end
 end
 
-function maximal_subgroup_reps(G::PermGroup)
-  return Oscar._as_subgroups(G, GAP.Globals.MaximalSubgroupClassReps(G.X))
-end
-
-"""
-An ascending chain of minimual supergroups linking `U` and `G`.
-Not a good algorithm, but Max' version `RefinedChain` does not produce
-minimal supergroups, ie. it fails in ``C_4``.
-"""
-function maximal_subgroup_chain(G::PermGroup, U::PermGroup)
-  l = [G]
-  while order(l[end]) > order(U)
-    m = maximal_subgroups(l[end])
-    push!(l, m[findfirst(x -> issubgroup(x, U)[1], m)])
-  end
-  return reverse(l)
-
-  l = GAP.Globals.AscendingChain(G.X,U.X)
-  map(GAP.Globals.MaximalSubgroupClassReps, l)
-  ll = GAP.Globals.RefinedChain(G.X,l)
-  return [Oscar._as_subgroup(G, x)[1] for x = ll]
-end
-
-function transitive_group_identification(G::PermGroup)
-  if degree(G) > 31
-    return -1
-  end
-  return GAP.Globals.TransitiveIdentification(G.X)
-end
-
 #TODO:
 #- ansehen der ZykelTypen um Sn/An zu erkennen
 #- Bessere Abstraktion um mehr Grundkoerper/ Ringe zu erlauben
 #- Bessere Teilkpoerper: ich brauche "nur" maximale
 #- sanity-checks
 #- "datenbank" fuer Beispiele
-
-# TODO: add a GSet Julia type which does something similar Magma's,
-# or also to GAP's ExternalSet (but don't use ExternalSet, to avoid the overhead)
-
-function all_blocks(G::PermGroup)
-  # TODO: this returns an array of arrays;
-  # TODO: AllBlocks assumes that we act on MovedPoints(G), which
-  # may NOT be what we want...
-  return GAP.gap_to_julia(Vector{Vector{Int}}, GAP.Globals.AllBlocks(G.X))
-end
-
-# TODO: update stabilizer to use GSet
-# TODO: allow specifying actions other than the default
-function stabilizer(G::Oscar.GAPGroup, seed, act)
-    return Oscar._as_subgroup(G, GAP.Globals.Stabilizer(G.X, GAP.julia_to_gap(seed), act))
-end
-
-# TODO: add type BlockSystem
-
-# TODO: perhaps get rid of set_stabilizer again, once we have proper Gsets
-function set_stabilizer(G::Oscar.GAPGroup, seed::Vector{Int})
-    return stabilizer(G, GAP.julia_to_gap(seed), GAP.Globals.OnSets)
-end
-
-# TODO: add lots of more orbit related stuff
-
-#provided by Thomas Breuer:
-
-Hecke.orbit(G::PermGroup, i::Int) = GAP.gap_to_julia(GAP.Globals.Orbit(G.X, GAP.julia_to_gap(i)))
-orbits(G::PermGroup) = GAP.gap_to_julia(GAP.Globals.Orbits(G.X))
-
-function action_homomorphism(G::PermGroup, omega::AbstractVector{Int})
-  mp = GAP.Globals.ActionHomomorphism(G.X, GAP.julia_to_gap(omega))
-  if mp == GAP.Globals.fail throw(ArgumentError("Invalid input")) end
-  H = PermGroup(GAP.Globals.Range(mp))
-  return GAPGroupHomomorphism{typeof(G), typeof(H)}(G, H, mp)
-end
-
-function block_system(G::PermGroup, B::Vector{Int})
-  orb = GAP.Globals.Orbit(G.X, GAP.julia_to_gap(B), GAP.Globals.OnSets)
-  GAP.gap_to_julia(Vector{Vector{Int}}, orb)
-end
-
-# given a perm group G and a block B, compute a homomorphism into Sym(B^G)
-function action_on_blocks(G::PermGroup, B::Vector{Int})
-  orb = GAP.Globals.Orbit(G.X, GAP.julia_to_gap(B), GAP.Globals.OnSets)
-  act = GAP.Globals.ActionHomomorphism(G.X, orb, GAP.Globals.OnSets)
-  H = GAP.Globals.Image(act)
-  T = Oscar._get_type(H)
-  H = T(H)
-  return Oscar._hom_from_gap_map(G, H, act)
-end
-
-Base.sign(G::PermGroup) = GAP.Globals.SignPermGroup(G.X)
-
-Base.isodd(G::PermGroup) = sign(G) == -1
-Base.iseven(n::PermGroup) = !isodd(n)
 
 @doc Markdown.doc"""
     to_elementary_symmetric(f)
@@ -583,45 +431,6 @@ function isprobably_invariant(g, p)
   n = ngens(R)
   lp = [rand(k) for i=1:n]
   return evaluate(g, lp) == evaluate(g^p, lp)
-end
-
-@doc Markdown.doc"""
-    short_right_transversal(G::PermGroup, H::PermGroup, s) ->
-
-Determines representatives for all right-cosets of `G` modulo `U`
-containing the element `s`.
-"""
-function short_right_transversal(G::PermGroup, H::PermGroup, s)
-  C = GAP.Globals.ConjugacyClasses(H.X)
-  cs = GAP.Globals.CycleStructurePerm(s.X)
-  can = []
-  for i=1:length(C)
-    c = C[i]
-    r = GAP.Globals.Representative(c)
-    if cs == GAP.Globals.CycleStructurePerm(r)
-      push!(can, r)
-    end
-  end
-
-  R = []
-  for c = can
-    d = GAP.Globals.RepresentativeAction(G.X, c, s.X)
-    if d != GAP.Globals.fail
-      push!(R, Oscar.group_element(G, d))
-      @assert Oscar.group_element(G, c)^R[end] == s
-    end
-  end
-
-  S = []
-  C = centralizer(G, s)[1]
-  for r = R
-    CH = centralizer(H^r, s)[1]
-    for t = right_transversal(C, CH)
-      push!(S, r*t)
-    end
-  end
-
-  return S
 end
 
 @doc Markdown.doc"""
@@ -1254,140 +1063,8 @@ function fixed_field(GC::GaloisCtx, U::PermGroup, extra::Int = 5)
   return number_field(Hecke.power_sums_to_polynomial(ps), check = false, cached = false)[1]
 end
 
-#TODO: do this properly, with indexed types and such
-# Partially Ordered Set ...
-mutable struct POSet{T}
-  elem::Array{T, 1}
-  "for two elements that are comparable, return -1, 0, 1"
-  cmp::Function #(::T, ::T) -> -1, 0, 1
-  "take two elems and test if they are comparable, returns Bool"
-  can_cmp::Function #(::T, ::T) -> Bool
-
-  function POSet{S}(elem::Array{S, 1}, can_cmp::Function, cmp::Function) where {S}
-    r = new{S}()
-    r.elem = elem
-    r.cmp = cmp
-    r.can_cmp = can_cmp
-    return r
-  end
-end
-
-POSet(elem::Array{S, 1}, can_cmp::Function, cmp::Function) where {S} = POSet{S}(elem, can_cmp, cmp)
-
-struct POSetElem{T}
-  e::Int
-  p::POSet{T}
-  function POSetElem(L::POSet{S}, i::Int) where {S}
-    return new{S}(i, L)
-  end
-  function POSetElem(L::POSet{S}, e::Any) where {S}
-    return new{S}(findfirst(x->L>can_cmp(e, L.elem[i]) && L.cmp(e, L.elem[i]) == 0, 1:length(L.elem)), L)
-  end
-end
-
-function maximal_elements(L::POSet)
-  d = Dict{typeof(L.elem[1]), Array{Int, 1}}()
-  for i = 1:length(L.elem)
-    e = L.elem[i]
-    d[e] = Int[]
-    for j=1:length(L.elem)
-      if i == j
-        continue
-      end
-      if L.can_cmp(e, L.elem[j]) && L.cmp(e, L.elem[j]) > 0
-        push!(d[e], j)
-      end
-    end
-  end
-  return [x for (x,v) = d if length(v) == 0]
-end
-
-function minimal_elements(L::POSet)
-  d = Dict{typeof(L.elem[1]), Array{Int, 1}}()
-  for i = 1:length(L.elem)
-    e = L.elem[i]
-    d[e] = Int[]
-    for j=1:length(L.elem)
-      if i == j
-        continue
-      end
-      if L.can_cmp(e, L.elem[j]) && (L.cmp(e, L.elem[j]) < 0)
-        push!(d[e], j)
-      end
-    end
-  end
-  return [x for (x,v) = d if length(v) == 0]
-end
-
-
-function Base.getindex(S::Hecke.SubSetSizeItr, i::Int)
-  return Hecke.int_to_elt(S, i)
-end
-
-struct BlockSystems
-  n::Int
-  l::Int
-  cur::Array{Array{Int, 1}, 1}
-  function BlockSystems(n::Int, l::Int)
-    @assert n % l == 0
-    return new(n, l, [collect((i-1)*l+1:i*l) for i=1:divexact(n, l)])
-  end
-end
-
-function Base.iterate(B::BlockSystems)
-  return B.cur, deepcopy(B.cur)
-end
-
-function Base.iterate(B::BlockSystems, st::Array{Array{Int, 1}})
-  if B.l==1||B.l==B.n
-    return nothing
-  end
-  i = length(B.cur)-1
-  while true
-    j = B.l
-    while true
-      if st[i][j] < B.n - B.l + j
-        st[i][j] += 1
-        free = Set(1:B.n)
-        for l=1:i-1
-          setdiff!(free, st[l])
-        end
-        if !(st[i][j] in free) 
-          continue
-        end
-        if length(intersect(free, Set(st[i][j]+1:B.n)))<B.l-j
-          continue
-        end
-        setdiff!(free, st[i][1:j])
-        while j < B.l
-          j += 1
-          I = intersect(free, Set(st[i][j-1]:B.n))
-          if isempty(I)
-            break
-          end
-          st[i][j] = minimum(I)
-          pop!(free, st[i][j])
-        end
-        i += 1
-        while i <= length(st)
-          for j=1:B.l
-            st[i][j] = minimum(free)
-            pop!(free, st[i][j])
-          end
-          i += 1
-        end
-        return deepcopy(st), st
-      end
-      j -= 1
-      if j == 1
-        i -= 1
-        i == 0 && return nothing
-        break
-      end
-    end
-  end
-end
-Base.IteratorSize(::BlockSystems) = Base.SizeUnknown()
+include("Group.jl")
+include("POSet.jl")
 
 end
